# perf: Improve generator perf