--- a/packages/babel-generator/src/node/whitespace.ts
+++ b/packages/babel-generator/src/node/whitespace.ts
@@ -14,6 +14,7 @@ import {
   isOptionalMemberExpression,
   isStringLiteral,
 } from "@babel/types";
+import * as charCodes from "charcodes";
 
 import type { NodeHandlers } from "./index";
 
@@ -76,7 +77,10 @@ function isHelper(node: t.Node): boolean {
   if (isMemberExpression(node)) {
     return isHelper(node.object) || isHelper(node.property);
   } else if (isIdentifier(node)) {
-    return node.name === "require" || node.name.charCodeAt(0) === 95; //node.name[0] === "_";
+    return (
+      node.name === "require" ||
+      node.name.charCodeAt(0) === charCodes.underscore
+    );
   } else if (isCallExpression(node)) {
     return isHelper(node.callee);
   } else if (isBinary(node) || isAssignmentExpression(node)) {
@@ -114,6 +118,7 @@ export const nodes: NodeHandlers<WhitespaceFlag> = {
         ? WhitespaceFlag.before | WhitespaceFlag.after
         : WhitespaceFlag.after;
     }
+    return 0;
   },
 
   /**
@@ -139,6 +144,7 @@ export const nodes: NodeHandlers<WhitespaceFlag> = {
     if (isFunction(node.left) || isFunction(node.right)) {
       return WhitespaceFlag.after;
     }
+    return 0;
   },
 
   /**
@@ -149,6 +155,7 @@ export const nodes: NodeHandlers<WhitespaceFlag> = {
     if (isStringLiteral(node) && node.value === "use strict") {
       return WhitespaceFlag.after;
     }
+    return 0;
   },
 
   /**
@@ -159,12 +166,14 @@ export const nodes: NodeHandlers<WhitespaceFlag> = {
     if (isFunction(node.callee) || isHelper(node)) {
       return WhitespaceFlag.before | WhitespaceFlag.after;
     }
+    return 0;
   },
 
   OptionalCallExpression(node: t.OptionalCallExpression): WhitespaceFlag {
     if (isFunction(node.callee)) {
       return WhitespaceFlag.before | WhitespaceFlag.after;
     }
+    return 0;
   },
 
   /**
@@ -185,6 +194,7 @@ export const nodes: NodeHandlers<WhitespaceFlag> = {
         return WhitespaceFlag.before | WhitespaceFlag.after;
       }
     }
+    return 0;
   },
 
   /**
@@ -195,6 +205,7 @@ export const nodes: NodeHandlers<WhitespaceFlag> = {
     if (isBlockStatement(node.consequent)) {
       return WhitespaceFlag.before | WhitespaceFlag.after;
     }
+    return 0;
   },
 };
 
@@ -212,6 +223,7 @@ nodes.ObjectProperty =
       if (parent.properties[0] === node) {
         return WhitespaceFlag.before;
       }
+      return 0;
     };
 
 nodes.ObjectTypeCallProperty = function (
@@ -221,6 +233,7 @@ nodes.ObjectTypeCallProperty = function (
   if (parent.callProperties[0] === node && !parent.properties?.length) {
     return WhitespaceFlag.before;
   }
+  return 0;
 };
 
 nodes.ObjectTypeIndexer = function (
@@ -234,6 +247,7 @@ nodes.ObjectTypeIndexer = function (
   ) {
     return WhitespaceFlag.before;
   }
+  return 0;
 };
 
 nodes.ObjectTypeInternalSlot = function (
@@ -248,6 +262,7 @@ nodes.ObjectTypeInternalSlot = function (
   ) {
     return WhitespaceFlag.before;
   }
+  return 0;
 };
 
 /**