… benchmark/babel-generator/real-case/jquery.mjs
… packages/babel-generator/src/node/parentheses.ts
… scripts/integration-tests/utils/local-registry.sh