<!--
Before making a PR, please read our contributing guidelines
https://github.com/babel/babel/blob/main/CONTRIBUTING.md

Please note that the Babel Team requires two approvals before merging most PRs.

For issue references: Add a comma-separated list of a [closing word](https://help.github.com/articles/closing-issues-via-commit-messages/) followed by the ticket number fixed by the PR. (it should be underlined in the preview if done correctly)

If you are making a change that should have a docs update: submit another PR to https://github.com/babel/website
-->

| Q                        | A <!--(Can use an emoji 👍) -->
| ------------------------ | ---
| License                  | MIT

<!-- Describe your changes below in as much detail as possible -->

I love this PR, and I'm sure you do too!

This PR improves the performance of the generator by 100%, which is equivalent to a 50% reduction in time consumption.

Benchmark results are from the real world (jquery 3.6).
The running environment is nodejs 18.3 and Windows 11.


```
current 1 jquery 3.6: 45.99 ops/sec ±0.83% (22ms)
current 4 jquery 3.6: 11.31 ops/sec ±0.44% (88ms)
current 16 jquery 3.6: 2.74 ops/sec ±0.52% (364ms)
current 64 jquery 3.6: 0.67 ops/sec ±0.34% (1490ms)
baseline 1 jquery 3.6: 25.99 ops/sec ±3.81% (38ms)
baseline 4 jquery 3.6: 5.57 ops/sec ±0.88% (180ms)
baseline 16 jquery 3.6: 1.33 ops/sec ±2.17% (750ms)
baseline 64 jquery 3.6: 0.32 ops/sec ±6.15% (3090ms)
baseline 1 jquery 3.6: 26.15 ops/sec ±0.65% (38ms)
baseline 4 jquery 3.6: 5.55 ops/sec ±0.85% (180ms)
baseline 16 jquery 3.6: 1.3 ops/sec ±3.79% (770ms)
baseline 64 jquery 3.6: 0.33 ops/sec ±2.08% (3013ms)
current 1 jquery 3.6: 45.51 ops/sec ±0.38% (22ms)
current 4 jquery 3.6: 11.19 ops/sec ±0.49% (89ms)
current 16 jquery 3.6: 2.71 ops/sec ±0.38% (369ms)
current 64 jquery 3.6: 0.67 ops/sec ±0.29% (1495ms)
```


I made the following changes.

Significant performance improvements:

1. Optimize the string merging logic. When merging into a large string, it seems that some strange operations will greatly improve the performance. It is difficult to explain clearly with words. You can directly view the code.

2. Optimize the queue of printers to avoid frequent creation and destruction of objects.

3. Modify the row and column calculation logic to avoid searching for newlines in most cases.
Note: Now we need to add a parameter to check for newlines on elements that may wrap, but this is rare, I only found two in the current code.

4. In `parentheses.ts` and `whitespace.ts`, change the objects to bitwise operations. And add bounds checking to avoid reading elements outside the array and useless type judgment.


Minor performance improvements:

5. Add the `tokenChar` method, which can save some checks, and automatically modify the single-character text to the character code and call `tokenChar` through the compile-time plugin.

6. Treat the queue elements uniformly as a single character or the same character.
Note: There is an externally observable change here.
That is, we assume that indentation characters must contain only one character.
E.g:
"space+space", or "tab+tab"
instead of "space+tab".


There are also some minor modifications that may not have been mentioned, but that shouldn't matter.


It's worth mentioning that none of the tests were modified and all passed, so this should be safe overall.


What might be possible in the future:

1. Current bit operations cannot be optimized by constant folding, this should have little impact, but we can improve it in the future.

2. We can inline `types.isXXX` with a plugin, I don't know how much performance improvement this will bring, but maybe try.


<a href="https://gitpod.io/#https://github.com/babel/babel/pull/14701"><img src="https://gitpod.io/button/open-in-gitpod.svg"/></a>