@liuxingbaoyu 2022-07-08
https://github.com/babel/babel/pull/14701/commits/4ec0304c9f8ef5935f753fd922568e9fa98d6dab

Ah, so sorry!
I made a simple change to the master branch for CI testing in my local repo and it was mistakenly included in this PR when rebasing.😭

https://github.com/babel/babel/pull/14574