@nicolo-ribaudo 2022-07-08
Whops, I didn't notice it :sweat_smile: 

But CI still works and as you pointed out the running time isn't affected by much; that line was causing problems with yarn 3, so... let's keep it :shrug: